chore(storage): split upload boundaries
